requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
playwright==1.40.0
streamlit==1.28.2
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2